fix:위젯으로 분리
